+++ model/model_security.go
@@ -22,7 +22,7 @@ type Security struct {
 	// WKN
 	Wkn string `json:"wkn,omitempty"`
 	// Quote
-	Quote float32 `json:"quote,omitempty"`
+	Quote float64 `json:"quote,omitempty"`
 	// Currency of quote
 	QuoteCurrency string `json:"quoteCurrency,omitempty"`
 	// Type of quote. 'PERC' if quote is a percentage value, 'ACTU' if quote is the actual amount
@@ -30,17 +30,17 @@ type Security struct {
 	// Quote date in the format 'YYYY-MM-DD HH:MM:SS.SSS' (german time).
 	QuoteDate string `json:"quoteDate,omitempty"`
 	// Value of quantity or nominal
-	QuantityNominal float32 `json:"quantityNominal,omitempty"`
+	QuantityNominal float64 `json:"quantityNominal,omitempty"`
 	// Type of quantity or nominal value. 'UNIT' if value is a quantity, 'FAMT' if value is the nominal amount
 	QuantityNominalType string `json:"quantityNominalType,omitempty"`
 	// Market value
-	MarketValue float32 `json:"marketValue,omitempty"`
+	MarketValue float64 `json:"marketValue,omitempty"`
 	// Currency of market value
 	MarketValueCurrency string `json:"marketValueCurrency,omitempty"`
 	// Entry quote
-	EntryQuote float32 `json:"entryQuote,omitempty"`
+	EntryQuote float64 `json:"entryQuote,omitempty"`
 	// Currency of entry quote
 	EntryQuoteCurrency string `json:"entryQuoteCurrency,omitempty"`
 	// Current profit or loss
-	ProfitOrLoss float32 `json:"profitOrLoss,omitempty"`
+	ProfitOrLoss float64 `json:"profitOrLoss,omitempty"`
 }
